+++ Solidner/Views/Planning/Weekly/WeeklyPlanningView.swift
@@ -8,7 +8,11 @@
 import SwiftUI
 
 struct WeeklyPlanningView: View {
-    let mealPlans: [MealPlan] = MealPlan.mockMealsZero
+    let mealPlans: [MealPlan]
+
+    init(mealPlans: [MealPlan] = MealPlan.mockMealsZero) {
+        self.mealPlans = mealPlans
+    }
 
     var body: some View {
         VStack {
@@ -124,72 +128,6 @@ struct WeeklyPlanningView: View {
 }
 
 
-extension Date {
-    static func date(year: Int, month: Int, day: Int) -> Date? {
-        var dateComponents = DateComponents()
-        dateComponents.year = year
-        dateComponents.month = month
-        dateComponents.day = day
-
-        return Calendar.current.date(from: dateComponents)
-    }
-
-    static func - (lhs: Date, rhs: Int) -> Date {
-        return Calendar.current.date(byAdding: .day, value: -rhs, to: lhs)!
-    }
-
-    static func componentsBetweenDates(from startDate: Date, to endDate: Date) -> DateComponents {
-        Calendar.current.dateComponents([.day, .hour, .minute], from: startDate, to: endDate)
-    }
-
-    static func range(from startDate: Date, to endDate: Date) -> [Date] {
-        let diff = componentsBetweenDates(from: startDate, to: endDate).day!
-        return (0...diff).compactMap {
-            Calendar.current.date(byAdding: .day, value: $0, to: startDate)
-        }
-    }
-
-    var day: Int {
-        return Calendar.current.component(.day, from: self)
-    }
-
-    var weekday: Int {
-        return Calendar.current.component(.weekday, from: self)
-    }
-
-    var weekDayKor: String {
-        Weekday(rawValue: weekday)!.description
-    }
-
-    enum Weekday: Int {
-        case sunday = 1
-        case monday = 2
-        case tuesday = 3
-        case wednesday = 4
-        case thursday = 5
-        case friday = 6
-        case saturday = 7
-
-        var description: String {
-            switch self {
-            case .sunday:
-                return "일"
-            case .monday:
-                return "월"
-            case .tuesday:
-                return "화"
-            case .wednesday:
-                return "수"
-            case .thursday:
-                return "목"
-            case .friday:
-                return "금"
-            case .saturday:
-                return "토"
-            }
-        }
-    }
-}
 
 extension WeeklyPlanningView {
     // MARK: - Mock Meals
